+++ src/main/java/nextstep/auth/application/dto/GithubAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
