+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -6,7 +6,7 @@ import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.domain.FavoriteRepository;
-import nextstep.member.AuthenticationException;
+import nextstep.auth.AuthenticationException;
 import nextstep.member.application.MemberService;
 import nextstep.subway.applicaion.SectionService;
 import nextstep.subway.applicaion.StationService;
