+++ src/main/java/nextstep/favorite/ui/FavoriteController.java
@@ -5,7 +5,7 @@ import nextstep.favorite.application.dto.FavoriteCreateResponse;
 import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.member.domain.LoginMember;
-import nextstep.member.ui.AuthenticationPrincipal;
+import nextstep.auth.ui.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
