+++ src/main/java/nextstep/member/application/TokenService.java
@@ -1,8 +1,10 @@
 package nextstep.member.application;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.member.AuthenticationException;
-import nextstep.member.application.dto.TokenResponse;
+import nextstep.auth.AuthenticationException;
+import nextstep.auth.application.GithubClient;
+import nextstep.auth.application.JwtTokenProvider;
+import nextstep.auth.application.dto.TokenResponse;
 import nextstep.member.domain.Member;
 import org.springframework.stereotype.Service;
