+++ src/main/java/nextstep/subway/ui/ControllerExceptionHandler.java
@@ -2,7 +2,7 @@ package nextstep.subway.ui;
 
 import javax.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
-import nextstep.member.AuthenticationException;
+import nextstep.auth.AuthenticationException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
