+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.member.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.member.application.dto.GithubLoginRequest;
+import nextstep.auth.application.dto.GithubLoginRequest;
 import nextstep.member.application.dto.MemberResponse;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
