+++ src/test/java/nextstep/member/application/GithubClientTest.java
@@ -2,6 +2,7 @@ package nextstep.member.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.auth.application.GithubClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
